+++ src/components/@layout/SideBar/SideBarHeader/index.tsx
@@ -3,13 +3,14 @@ import { Suspense, useEffect, useState } from 'react';
 import Notification from '@/components/@common/SVG/Icon/Notification';
 import SideBarProfile from '../SideBarProfile';
 import SideBarProfileSkeleton from '../SideBarProfileSkeleton';
-import { useRecoilValue } from 'recoil';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 import isLoggedInAtom from '@/recoil/auth/isLoggedIn';
 import Plus from '@/components/@common/SVG/Icon/Plus';
 import { useNavigate } from 'react-router';
 import { CHAT_PATH } from '@/constants/routes';
 import useGetUnreadAlarmCount from '@/hooks/user/useGetUnreadAlarmCount';
-import AlarmPopover from '@/components/user/AlarmPopover';
+import AlarmPopover from '@/components/user/Alarm/AlarmPopover';
+import alarmListAtom from '@/recoil/user/alarmList';
 
 const SideBarHeader = () => {
   const navigate = useNavigate();
@@ -22,6 +23,8 @@ const SideBarHeader = () => {
 
   const { data: unreadAlarmCount } = useGetUnreadAlarmCount();
 
+  const setAlarmList = useSetRecoilState(alarmListAtom);
+
   useEffect(() => {
     if (unreadAlarmCount && unreadAlarmCount > 0) {
       setHasAlarm(true);
@@ -32,9 +35,15 @@ const SideBarHeader = () => {
     navigate(CHAT_PATH.CHAT_ROOM_CREATE);
   };
 
-  const handleOpenAlarm = () => {
+  const handleToggleAlarm = () => {
     setIsAlarmOpen((prevState) => !prevState);
   };
+
+  useEffect(() => {
+    if (!isAlarmOpen) {
+      setAlarmList([]);
+    }
+  }, [isAlarmOpen, setAlarmList]);
 
   return (
     <Styled.Wrapper>
@@ -47,7 +56,7 @@ const SideBarHeader = () => {
             <Plus title="채팅방 생성" width="32" height="32" />
           </Styled.SideBarButton>
           {isAlarmOpen && <AlarmPopover />}
-          <Styled.SideBarButton type="button" onClick={handleOpenAlarm}>
+          <Styled.SideBarButton type="button" onClick={handleToggleAlarm}>
             {hasAlarm && <Styled.NotificationBadge />}
             <Notification title="알림" width="32" height="32" />
           </Styled.SideBarButton>
