+++ src/components/user/Alarm/AlarmItem/AlarmItem.styled.ts
@@ -1,34 +1,4 @@
-import AlarmTriangle from '@/components/@common/SVG/Icon/AlarmTriangle';
 import styled from 'styled-components';
-
-export const AlarmPopoverContainer = styled.div`
-  position: absolute;
-  top: 64px;
-  left: -2px;
-  background-color: ${({ theme }) => theme.colors.White};
-  width: 428px;
-  height: 548px;
-  border-radius: 12px;
-  z-index: 100;
-  box-shadow: 0px 0px 12px 0px rgba(0, 0, 0, 0.1);
-  display: flex;
-  flex-direction: column;
-`;
-
-export const AlarmTriangleIcon = styled(AlarmTriangle)`
-  position: absolute;
-  top: -20px;
-  left: 43px;
-`;
-
-export const Header = styled.h3`
-  ${({ theme }) => theme.typo.T2_R}
-  color: ${({ theme }) => theme.colors.Gray_1000};
-  padding: 16px 20px;
-  text-align: left;
-`;
-
-export const AlarmList = styled.ul``;
 
 export const AlarmItem = styled.li`
   padding: 16px 20px;
@@ -95,14 +65,4 @@ export const AcceptButton = styled.button`
   border-radius: 4px;
   padding: 6px 12px;
   line-height: 21px;
-`;
-
-export const EmptyText = styled.p`
-  ${({ theme }) => theme.typo.B1_R};
-  color: ${({ theme }) => theme.colors.Gray_700};
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex: 1;
-  margin-bottom: 62px;
 `;
