--- src/components/user/AlarmPopover/index.tsx
@@ -1,152 +1,0 @@
-import useGetAlarmList from '@/hooks/user/useGetAlarmList';
-import * as Styled from './AlarmPopover.styled';
-import ProfileImage from '@/components/@common/ProfileImage';
-import { useEffect, useState } from 'react';
-import { AlarmType } from '@/types/user';
-import useRejectFriendRequest from '@/hooks/user/useRejectFriendRequest';
-import useAcceptFriendRequest from '@/hooks/user/useAcceptFriendRequest';
-import useAcceptChatRoomInvitation from '@/hooks/chat/useAcceptChatRoomInvitation';
-import useRejectChatRoomInvitation from '@/hooks/chat/useRejectChatRoomInvitation';
-import { useRecoilState } from 'recoil';
-import alarmListAtom from '@/recoil/user/alarmList';
-import ProfileDialog from '@/components/@common/Modal/ProfileDialog';
-import { IUserProfile } from '@/types/@common';
-import useGetProfile from '@/hooks/@common/useGetProfile';
-import { formatTimeAgo } from '@/utils/date';
-
-const AlarmPopover = () => {
-  const { data: alarmListResponse } = useGetAlarmList();
-
-  const [alarmList, setAlarmList] = useRecoilState(alarmListAtom);
-
-  const { mutate: acceptFriendRequest } = useAcceptFriendRequest();
-
-  const { mutate: rejectFriendRequest } = useRejectFriendRequest();
-
-  const { mutate: acceptChatRoomInvitation } = useRejectChatRoomInvitation();
-
-  const { mutate: rejectChatRoomInvitation } = useAcceptChatRoomInvitation();
-
-  const [isOpenProfile, setIsOpenProfile] = useState(false);
-
-  const [selectedProfileId, setSelectedProfileId] = useState<number | null>(
-    null,
-  );
-
-  const [selectedProfile, setSelectedProfile] = useState<IUserProfile | null>(
-    null,
-  );
-
-  const { data: userProfile } = useGetProfile(selectedProfileId);
-
-  const handleAccept = (type: AlarmType, alarmId: number) => {
-    if (type === AlarmType.FRIEND_REQUEST) {
-      acceptFriendRequest(alarmId);
-    }
-
-    if (type === AlarmType.CHAT_ROOM_INVITATION) {
-      acceptChatRoomInvitation(alarmId);
-    }
-  };
-
-  const handleReject = (type: AlarmType, alarmId: number) => {
-    if (type === AlarmType.FRIEND_REQUEST) {
-      rejectFriendRequest(alarmId);
-    }
-
-    if (type === AlarmType.CHAT_ROOM_INVITATION) {
-      rejectChatRoomInvitation(alarmId);
-    }
-  };
-
-  useEffect(() => {
-    if (alarmListResponse) {
-      setAlarmList(alarmListResponse.content);
-    }
-  }, [alarmListResponse, setAlarmList]);
-
-  useEffect(() => {
-    if (userProfile) {
-      setSelectedProfile(userProfile);
-    }
-  }, [userProfile]);
-
-  const handleOpenProfile = (memberId: number) => {
-    setSelectedProfileId(memberId);
-    setIsOpenProfile(true);
-  };
-
-  const handleCloseProfile = () => {
-    setSelectedProfileId(null);
-    setSelectedProfile(null);
-    setIsOpenProfile(false);
-  };
-
-  return (
-    <>
-      {isOpenProfile && selectedProfile && (
-        <ProfileDialog profile={selectedProfile} onClose={handleCloseProfile} />
-      )}
-      <Styled.AlarmPopoverContainer>
-        <Styled.AlarmTriangleIcon title="알림" width="25" height="20" />
-        <Styled.Header>전체 알림</Styled.Header>
-        {alarmList.length > 0 ? (
-          <Styled.AlarmList>
-            {alarmList.map((alarm) => (
-              <Styled.AlarmItem key={alarm.id}>
-                <Styled.Time>{formatTimeAgo(alarm.createdAt)}</Styled.Time>
-                <Styled.ContentContainer>
-                  {alarm.type === AlarmType.FRIEND_REQUEST ? (
-                    <Styled.ShowProfileButton
-                      type="button"
-                      onClick={() => handleOpenProfile(alarm.senderId)}
-                    >
-                      <ProfileImage
-                        size={40}
-                        src={alarm.senderProfileImage}
-                        alt={alarm.nickname}
-                      />
-                    </Styled.ShowProfileButton>
-                  ) : (
-                    <ProfileImage
-                      size={40}
-                      src={alarm.senderProfileImage}
-                      alt={alarm.nickname}
-                    />
-                  )}
-
-                  <div>
-                    <Styled.Content>
-                      <Styled.Nickname>{alarm.nickname}</Styled.Nickname>
-                      {alarm.type === AlarmType.FRIEND_REQUEST
-                        ? '님이 친구 추가 요청을 보냈어요'
-                        : '에서 나를 초대했어요'}
-                    </Styled.Content>
-                    <Styled.ButtonContainer>
-                      <Styled.RejectButton
-                        type="button"
-                        onClick={() => handleReject(alarm.type, alarm.id)}
-                      >
-                        거절
-                      </Styled.RejectButton>
-                      <Styled.AcceptButton
-                        type="button"
-                        onClick={() => handleAccept(alarm.type, alarm.id)}
-                      >
-                        수락
-                      </Styled.AcceptButton>
-                    </Styled.ButtonContainer>
-                  </div>
-                </Styled.ContentContainer>
-              </Styled.AlarmItem>
-            ))}
-          </Styled.AlarmList>
-        ) : (
-          <Styled.EmptyText>받은 알림이 없어요</Styled.EmptyText>
-        )}
-      </Styled.AlarmPopoverContainer>
-    </>
-  );
-};
-
-export default AlarmPopover;
